Move source code into `lib` folder
Also Use `files` in package.json
