+++ lib/formats.js
