+++ lib/require.js
@@ -1,6 +1,6 @@
 var fs = require('fs')
 var path = require('path')
-var compile = require('./')
+var compile = require('../')
 
 delete require.cache[require.resolve(__filename)]
